LayoutTests/platform/ios-sim-deprecated/svg/animations/animate-elem-02-t-drt-expected.txt

SVG 1.1 dynamic animation tests

A copy of the corresponding W3C-SVG-1.1 test, which dumps the animation at certain times

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect1.height.animVal.value is 200
PASS rect1.height.baseVal.value is 20
PASS rect2.height.animVal.value is 220
PASS rect2.height.baseVal.value is 20
PASS rect3.height.animVal.value is 200
PASS rect3.height.baseVal.value is 20
PASS rect1.height.animVal.value is 200
PASS rect1.height.baseVal.value is 20
PASS rect2.height.animVal.value is 220
PASS rect2.height.baseVal.value is 20
PASS rect3.height.animVal.value is 200
PASS rect3.height.baseVal.value is 20
PASS rect1.height.animVal.value is 20
PASS rect1.height.baseVal.value is 20
PASS rect2.height.animVal.value is 40
PASS rect2.height.baseVal.value is 20
PASS rect3.height.animVal.value is 20
PASS rect3.height.baseVal.value is 20
PASS rect1.height.animVal.value is 20
PASS rect1.height.baseVal.value is 20
PASS rect2.height.animVal.value is 40
PASS rect2.height.baseVal.value is 20
PASS rect3.height.animVal.value is 20
PASS rect3.height.baseVal.value is 20
PASS rect1.height.animVal.value is 200
PASS rect1.height.baseVal.value is 20
PASS rect2.height.animVal.value is 220
PASS rect2.height.baseVal.value is 20
PASS rect3.height.animVal.value is 220
PASS rect3.height.baseVal.value is 20
PASS rect1.height.animVal.value is 200
PASS rect1.height.baseVal.value is 20
PASS rect2.height.animVal.value is 220
PASS rect2.height.baseVal.value is 20
PASS rect3.height.animVal.value is 220
PASS rect3.height.baseVal.value is 20
PASS rect1.height.animVal.value is 20
PASS rect1.height.baseVal.value is 20
PASS rect2.height.animVal.value is 40
PASS rect2.height.baseVal.value is 20
PASS rect3.height.animVal.value is 40
PASS rect3.height.baseVal.value is 20
PASS rect1.height.animVal.value is 20
PASS rect1.height.baseVal.value is 20
PASS rect2.height.animVal.value is 40
PASS rect2.height.baseVal.value is 20
PASS rect3.height.animVal.value is 40
PASS rect3.height.baseVal.value is 20
PASS rect1.height.animVal.value is 20
PASS rect1.height.baseVal.value is 20
PASS rect2.height.animVal.value is 40
PASS rect2.height.baseVal.value is 20
PASS rect3.height.animVal.value is 40
PASS rect3.height.baseVal.value is 20
PASS rect1.height.animVal.value is 20
PASS rect1.height.baseVal.value is 20
PASS rect2.height.animVal.value is 40
PASS rect2.height.baseVal.value is 20
PASS rect3.height.animVal.value is 40
PASS rect3.height.baseVal.value is 20
PASS successfullyParsed is true

TEST COMPLETE
